+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.7.6-dev
  - properly subtract previous statistic when handling `set_failure()` and `set_success()`
  - detect and track redirects in `GooseRawRequest`
+ - `--sticky-follow` makes redirect of GooseClient base_url sticky, affecting subsequent requests
 
 ## 0.7.5 June 10, 2020
  - store actual URL requested in GooseRawRequest